Add more droptables

+++ src/me/neoblade298/neocore/shared/droptables/Droppable.java
@@ -1,19 +1,19 @@
 package me.neoblade298.neocore.shared.droptables;
 
 public class Droppable<E> {
-	private int weight;
+	private double weight;
 	private E drop;
 	
 	public Droppable(E drop) {
 		this.drop = drop;
 	}
 	
-	public Droppable(E drop, int weight) {
+	public Droppable(E drop, double weight) {
 		this.drop = drop;
 		this.weight = weight;
 	}
 	
-	public int getWeight() {
+	public double getWeight() {
 		return weight;
 	}
